V2 drm hisi provide canned 720p if no other mode
This creates a canned DRM mode for 720p60, and in the event there is
no usable mode left after DRM checks validity, the canned mode will be
used.

This is helpful in two cases

1) No EDID appeared, but the device supports the fallback mode

2) An EDID appeared, but it does not list any acceptable mode.  However,
the device does actually support the fallback mode.

It's implemented by adding a new optional connector func callback which
is called in the event we are left with no valid modes.  Connectors may
handle the "fallback_mode" callback and add a fallback mode to use rather
than the DRM default of 1024 x 768.

You can test "no EDID" by shorting HDMI connector p16 (SDA) to p17 (0V)

Signed-off-by: Andy Green <andy.green@linaro.org>