clk: hi6220: syspll is set by bootloader. Read from registers during driver init.
SYSPLL is set by bootloader. Hikey has two different bootloader rate settings due
to historical reasons. So, in order to compatible to both bootloader settings, in
kernel space, when clk init, we read the actual setting from registers.

Previously, this is assumed static and hardcoded in a static data structure.

Signed-off-by: Guodong Xu <guodong.xu@linaro.org>